Remove redundant local ba_schur.rs; consolidate on kornia_3d + vi_ba_schur
The pipeline's visual-only BA already runs kornia_3d::ba_schur::bundle_adjust_schur
(map.rs), so the local ba_schur.rs `bundle_adjust_schur`/`_with_priors` were a
third, unused copy exercised only by their own tests. Delete the module and move
the three reprojection helpers it hosted (pose_to_se3 / se3_to_pose /
residual_and_jacobians) into vi_ba_schur.rs, their only remaining user.

- delete crates/kornia-slam/src/ba_schur.rs (+ its `pub mod` in lib.rs)
- move the 3 helpers into vi_ba_schur.rs as private fns

Verified bit-exact vs develop on MH_01 (500 frames), mono+IMU and stereo+IMU:
raw + aligned trajectory CSVs identical, ATE/scale unchanged (mono 0.1168m/0.614,
stereo 0.0226m/0.997). kornia-slam tests: 28 pass (the 4 removed were ba_schur.rs's
own tests for the deleted redundant solver).
