+++ crates/kornia-slam/src/lib.rs
@@ -1,6 +1,5 @@
 //! Visual odometry and SLAM building blocks for kornia-rs.
 
-pub mod ba_schur;
 pub mod estimation;
 pub mod frame;
 pub mod map;
